Generate SystemVerilog interfaces with only unpacked arrays
