+++ cava/Cava2SystemVerilog.hs
@@ -1,4 +1,4 @@
-{- Copyright 2020 The Project Oak Authors
+{- Copyright 2021 The Project Oak Authors
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE ViewPatterns #-}
 
-module Cava2SystemVerilog
+module Cava2SystemVerilog (module Cava2SystemVerilog, module TestBench)
 where
 
 import Control.Monad.State.Lazy
@@ -29,6 +29,18 @@ import qualified BinNums
 import Netlist
 import Signal
 import qualified StateMonad
+
+import SystemVerilogUtils
+
+import TestBench
+
+{-
+NOTES.
+======
+The current SystemVerilog extraction process only generates packed arrays.
+All arrays are interpretated as having "downto" indexing e.g. [3:0].
+An attempt is made to unsmash literals to recover stems of arrays.
+-}
 
 writeSystemVerilog :: CavaState -> IO ()
 writeSystemVerilog cavastate@(Netlist.Coq_mkCavaState _ _ _ _
@@ -65,16 +77,10 @@ cava2SystemVerilog cavaState@(Netlist.Coq_mkCavaState netNumber vCount' vDefs' e
     ["  );"] ++
     ["",
      "  timeunit 1ns; timeprecision 1ns;",
-     "",
-     "  logic zero;",
-     "  logic one;"] ++
+     ""] ++
     declareLocalNets (fromN netNumber) ++
     [vectorDeclaration ("v" ++ show i) k s ++ ";" | ((k, s), i) <- zip vDefs [0..]] ++
     [ "  " ++ t ++ " ext_" ++ show i ++ ";" | (t, i) <- zip ext [0..]] ++
-    [""] ++
-    ["  // Constant nets",
-     "  assign zero = 1'b0;",
-     "  assign one = 1'b1;"] ++
     [""] ++
     [generateInstance genState inst i | (inst, i) <- zip instances [0..]] ++
     [""] ++
@@ -98,7 +104,7 @@ declareLocalNets n
   = if n == 0 then
       []
     else
-      ["  logic[" ++ show (n-1) ++ ":0] net;"]
+      ["  logic net[0:" ++ show (n-1) ++ "];"]
 
 data NetlistGenerationState
   = NetlistGenerationState (Maybe Signal) SignalEdge (Maybe Signal) SignalEdge
@@ -107,45 +113,29 @@ inputPorts :: [Netlist.PortDeclaration] -> [String]
 inputPorts = map inputPort
 
 inputPort :: Netlist.PortDeclaration -> String
-inputPort (Coq_mkPort name Bit) = "  input logic " ++ name
-inputPort (Coq_mkPort name (Vec k s))
-  = "  input " ++ vectorDeclaration name k s
-inputPort  (Coq_mkPort name (ExternalType typeName))
-  = "  input " ++ typeName ++ " " ++ name
-
-vectorDeclaration :: String -> SignalType -> Integer -> String
-vectorDeclaration name k s
-  = case k of
-      Bit -> "  logic[" ++ show (s - 1) ++ ":0] " ++ name
-      Vec k2 s2 -> vectorDeclaration name k2 s2 ++ "[" ++ show s ++ "]"
+inputPort (Coq_mkPort name typ)
+  = "  input " ++ declareSignal name typ
 
 outputPorts :: [Netlist.PortDeclaration] -> [String]
 outputPorts = map outputPort
 
 outputPort :: Netlist.PortDeclaration -> String
-outputPort (Netlist.Coq_mkPort name Bit) = "  output logic " ++ name
-outputPort (Netlist.Coq_mkPort name (Vec k s))
-  = "  output " ++ vectorDeclaration name k s
-outputPort  (Coq_mkPort name (ExternalType typeName))
-  = "  output " ++ typeName ++ " " ++ name
-
-insertCommas :: [String] -> [String]
-insertCommas [] = []
-insertCommas [x] = [x]
-insertCommas (x:y:xs) = (x ++ ",") : insertCommas (y:xs)
+outputPort (Netlist.Coq_mkPort name typ)
+  = "  output " ++ declareSignal name typ
 
 showVectorElements :: [Signal] -> String
 showVectorElements e
   = concat (insertCommas (map showSignal e))
 
+-- Show vector literals, interpretating them as packed arrays
+-- with downto indexing.
 showVecLiteral :: SignalType -> [Signal] -> String
 showVecLiteral k e
   = case k of
       Bit -> -- Packed vector, downto indexing
              "{" ++ showVectorElements (reverse e) ++ "}"
-      _   -> -- unpacked vector, upto indexing
-             "'{" ++ showVectorElements e ++ "}"
-
+      _   -> -- unpacked vector, downto indexing
+             "'{" ++ showVectorElements (reverse e) ++ "}"
 showSignal :: Signal -> String
 showSignal signal
   = case signal of
@@ -153,8 +143,8 @@ showSignal signal
       UninterpretedSignal _ name -> name
       UninterpretedSignalIndex _ i -> "ext_" ++ show (fromN i)
       SelectField _ _ sk1 f -> showSignal sk1 ++ "." ++ f
-      Gnd -> "zero"
-      Vcc -> "one"
+      Gnd -> "1'b0"
+      Vcc -> "1'b1"
       Wire n -> "net[" ++ show (fromN n) ++ "]"
       NamedWire name -> name
       NamedVector _ _ name -> name
@@ -278,238 +268,6 @@ showConstExpr constExpr =
   case constExpr of
     HexLiteral w v -> show w ++ "'h" ++ showHex (fromN v) ""
     StringLiteral s -> "\"" ++ s ++ "\""
-
---------------------------------------------------------------------------------
--- Generate test bench
---------------------------------------------------------------------------------
-
-writeTestBench :: TestBench -> IO ()
-writeTestBench testBench
-  = do putStr ("Generating test bench " ++ filename ++ " " ++ driver ++ "...")
-       writeFile filename (unlines (generateTestBench testBench))
-       writeFile driver (unlines (cppDriver name clk ticks))
-       writeFile (name ++ ".tcl") (unlines (tclScript name ticks))
-       putStrLn (" [done]")
-    where
-    name = testBenchName testBench
-    filename = name ++ ".sv"
-    driver = name ++ ".cpp"
-    ticks = length (testBenchInputs testBench)
-    clk' = clkName (testBenchInterface testBench)
-    clk = if clk' == "" then
-            "clk" -- This is a combinational circuit, use make up a clock
-                  -- name for the test bench ticks.
-          else
-            clk'
-
-generateTestBench :: TestBench -> [String]
-generateTestBench testBench
-  = ["// Automatically generated SystemVerilog 2012 code from Cava",
-     "// Please do not hand edit.",
-     "",
-     "module " ++ (testBenchName testBench) ++ "(",
-     "  input logic clk",
-     ");",
-     "",
-     "  timeunit 1ns; timeprecision 1ns;",
-     "",
-     "  logic rst;",
-     "",
-     "  " ++ name ++ " " ++ name ++ "_inst (.*);",
-     "",
-     "  // Circuit inputs"] ++
-     declareLocalPorts (circuitInputs intf) ++
-    ["  // Circuit outputs"] ++
-    declareLocalPorts (circuitOutputs intf) ++
-    [] ++
-    (if length (testBenchInputs testBench) == 0 then [] else body) ++
-    [] ++
-    ["endmodule"
-    ]
-    where
-    intf = testBenchInterface testBench
-    name = circuitName intf
-    inputPortList = circuitInputs intf
-    outputPortList = circuitOutputs intf
-    nrTests = length (testBenchInputs testBench)
-    body =
-      ["  // Input test vectors"] ++
-      initTestVectors inputPortList (testBenchInputs testBench) ++
-      ["",
-       "  // Expected output test vectors"] ++
-      initTestVectors outputPortList (testBenchExpectedOutputs testBench) ++
-      ["",
-      "  int unsigned i_cava = 0;"] ++
-      assignInputs inputPortList ++
-      ["",
-      "  always @(posedge clk) begin"
-      ] ++
-      [addDisplay (inputPortList ++ outputPortList)] ++
-      checkOutputs outputPortList ++
-      ["    if (i_cava == " ++ show (nrTests - 1) ++ ") begin",
-      "      $display (\"PASSED\");",
-      "      i_cava <= 0;",
-      "    end else begin",
-      "      i_cava <= i_cava + 1 ;",
-      "    end;",
-      "  end"
-      ]
-
-declareCircuitPort :: PortDeclaration -> String
-declareCircuitPort port
-  = "  output " ++ declarePort port
-
-declareLocalPorts ::  [PortDeclaration] -> [String]
-declareLocalPorts = map declareLocalPort 
-
-declareLocalPort :: PortDeclaration -> String
-declareLocalPort port
-  = declarePort port ++ ";"
-
-declarePort :: PortDeclaration -> String
-declarePort (Coq_mkPort name kind) =
-  case kind of
-    Bit -> "  (* mark_debug = \"true\" *) logic " ++ name
-    Vec k s -> "  (* mark_debug = \"true\" *) " ++ vectorDeclaration name k s
-
-initTestVectors :: [PortDeclaration] -> [[SignalExpr]] -> [String]
-initTestVectors _ [] = [] -- No test vectors provided, SystemVerilog does not permit zero length arrays
-initTestVectors [] _ = []
-initTestVectors (p:ps) s
-  = initTestVector p (map head s) ++
-    initTestVectors ps (map tail s)
-
-initTestVector :: PortDeclaration -> [SignalExpr] -> [String]
-initTestVector pd@(Coq_mkPort name typ) s
-  = [declarePort (Coq_mkPort name' typ) ++ " = '{"] ++
-    insertCommas (map showSignalExpr s) ++
-    ["    };"]
-    where
-    name' = name ++ "_vectors[" ++ show (length s) ++ "]"
-
-showSignalExpr :: SignalExpr -> String
-showSignalExpr (BitVal v)   = "    1'b" ++ showBit v
-showSignalExpr (VecVal xs) | isAllBits xs
-  = "    " ++ show (length xs) ++ "'d" ++ show (signalToInt xs)
-showSignalExpr (VecVal xs)
-  = "    '{ " ++ concat (insertCommas (map showSignalExpr xs )) ++ " }"
-
-isAllBits :: [SignalExpr] -> Bool
-isAllBits = and . map isBitVal
-
-isBitVal :: SignalExpr -> Bool
-isBitVal (BitVal _) = True
-isBitVal _ = False
-
-signalToInt :: [SignalExpr] -> Integer
-signalToInt [] = 0
-signalToInt ((BitVal b):xs)
-  = case b of
-      False -> 2 * signalToInt xs
-      True -> 1 + 2 * signalToInt xs
-signalToBit _ = error "Not a bit-vector"
-
-showBit :: Bool -> String
-showBit False = "0"
-showBit True = "1"
-
-assignInputs :: [PortDeclaration] -> [String]
-assignInputs = map assignInput
-
-assignInput :: PortDeclaration -> String
-assignInput port
-  = "  assign " ++ port_name port ++ " = " ++ port_name port ++ "_vectors[i_cava];"
-
-addDisplay ::  [PortDeclaration] -> String
-addDisplay ports
-  = "      $display(\"" ++ formatArgs ++ "\", $time, " ++ formatPars ++ ");"
-    where
-    formatArgs = concat (insertCommas
-                         ("%t: tick = %0d": map formatPortWithName ports))
-    formatPars = concat (insertCommas ("i_cava": map smashPorts ports))
-
-formatPortWithName :: PortDeclaration -> String
-formatPortWithName (Coq_mkPort name Bit) = " " ++ name ++ " = %0b"
-formatPortWithName (Coq_mkPort name (Vec Bit _)) = " " ++ name ++ " = %0d"
-formatPortWithName (Coq_mkPort name (Vec (Vec _ _) s))
-  = concat (insertCommas [" " ++ name ++ "[" ++ show i ++ "] = %0d" |
-            i  <- [0..s-1]])
-
-smashPorts :: PortDeclaration -> String
-smashPorts portDec
-  = case port_type portDec of
-      Bit -> name
-      Vec Bit _ -> name
-      Vec (Vec _ _) s ->
-        concat (insertCommas [name ++ "[" ++ show i ++ "]" | i <- [0..s-1]])
-    where
-    name = port_name portDec
-
-checkOutputs :: [PortDeclaration] -> [String]
-checkOutputs ports = concat (map checkOutput ports)
-
-checkOutput :: PortDeclaration -> [String]
-checkOutput port
-  = ["      if(" ++ name ++ " != " ++ name ++ "_vectors[i_cava]) begin",
-     "        $error (\"For " ++ name ++ " expected " ++ fmt ++ " but got " ++
-     fmt ++ "\", " ++ name ++ "_vectors[i_cava], " ++ name ++ ");",
-     "      end;"
-    ]
-    where
-    fmt = formatKind (port_type port)
-    name = port_name port
-
-formatKind :: SignalType -> String
-formatKind Bit = "%0b"
-formatKind (Vec Bit _) = "%0d"
-formatKind other = "error: attempt to format unknown kind"
-
-cppDriver :: String -> String -> Int -> [String]
-cppDriver name clkName ticks =
-  ["#include <stdlib.h>",
-    "#include \"V" ++ name ++ ".h\"",
-    "#include \"verilated.h\"",
-    "#include \"verilated_vpi.h\"",
-    "#include \"verilated_vcd_c.h\"",
-    "",
-    "double main_time = 0;",
-    "double sc_time_stamp() { return main_time; }",
-    "",
-    "int main(int argc, char **argv) {",
-    "  Verilated::commandArgs(argc, argv);",
-    "  V" ++ name ++ " *top = new V" ++ name ++ ";",
-    "  VerilatedVcdC* vcd_trace = new VerilatedVcdC;",
-    "  Verilated::traceEverOn(true);",
-    "  top->trace(vcd_trace, 99);",
-    "  top->eval(); vcd_trace->dump(main_time);",
-    "  vcd_trace->open(\"" ++ name ++ ".vcd\");",
-    "  for (unsigned int i = 0; i < " ++ show ticks ++ "; i++) {",
-    "    top->" ++ clkName ++ " = 0; main_time += 5;",
-    "    top->eval(); vcd_trace->dump(main_time);",
-    "    top->" ++ clkName ++ " = 1;  main_time += 5;",
-    "    top->eval(); vcd_trace->dump(main_time);",
-    "  }",
-    "  vcd_trace->close();",
-    "  delete top;",
-    "  exit(EXIT_SUCCESS);",
-    "}"
-  ]
-
--- This function generates a TCL script for driving the top-level
--- testbench generated from Cava for use with the Vivado xsim
--- simulator. The test is run for number of cycles that correspond to
--- the required number of ticks and a VCD wavefile file is generated.
-tclScript :: String -> Int -> [String]
-tclScript name ticks
-  = ["open_vcd " ++ name ++ ".vcd",
-     "log_vcd *",
-     "log_vcd [ get_objects * ]",
-     "add_force {/" ++ name ++ "/clk} {0 0ns} {1 50ns} -repeat_every 100ns",
-     "run " ++ show ticks ++ "00ns",
-     "flush_vcd",
-     "close_vcd",
-     "exit"
-    ]
 
 --------------------------------------------------------------------------------
 -- unsmashSignaling vector literals.
